shill: net: Avoid CHECK in IPAddress::CanReachAddress()
Currently, IPAddress::CanReachAddress() uses CHECK_EQ to check if the
two addresses have the same family, but this function is easy to be
misused for addresses of different families, and a crash will happen
in that case.

This patch changes it into a normal if statement. Also updates the
corresponding unit tests.

BUG=chromium:1126781
TEST=cros_workon_make --board=$BOARD --test shill

Change-Id: Ibc0b7bc2e3e46c234332779c54c2be5265c94461
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2402818
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>