lorgnette: Include IPP-USB devices in ListScanners
crrev.com/c/2404084 added the ippusb backend that will scan through an
IPP-USB tunnel, but the caller had to know how to find a scanner and
construct a device name referring to it.  This CL now adds enumeration
so that IPP-USB scanners can be listed like other sane backends.
Enumeration happens in two stages:

1. For every USB device that includes an interface with bInterfaceClass
of printer and bInterfaceProtocol of IPP-USB, generate an ippusb: device
name referring to this device.

2. For each string generated in step 1, call SaneClient::ConnectToDevice
to see if the device responds to eSCL on its IPP-USB tunnel.

Devices that pass the check in step 2 are added to the list of scanners
returned in ListScanners.

BUG=b:159484077
TEST=lorgnette_cli with several USB scanners attached

Cq-Depend: chromium:2414734
Change-Id: I1feb9b26f85cabd20047189335173eb808655e6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2417029
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>