ml: add tflite benchmark driver.
(1) Add benchmark_start and  free_benchmark_results ffor ml_benchmark to
run.

(2) To make shared_library, we have to add use_pic=true in several
places of mlservice and some libs it depends on. For the same reason, we
can't let benchmark_start depends on MachineLearningServiceImpl; if we
do that, we'll have to modify many places. So instead, we used ModelImpl
for the inference directly.

(3) A expected input-output pair is also required for both inferencing
and accuracy calculation.

(4) A ExpectedInputOutput::Example is added manually for now, which is
equivalent to tensorflow::Example for deserializing expected input. This
should be replaced with tensorflow::Example when the proto is available
in mlservice.

(5) unit tests are added to verify the inference runs successfully.

BUG=chromium:1107286
TEST='All unit tests passed.'

Cq-Depend: chromium:2411999
Change-Id: Iefb1f21819be07052efd4e28a988896883c9baaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2409847
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Charles . <charleszhao@chromium.org>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Michael Mortensen <mmortensen@google.com>
Tested-by: Charles . <charleszhao@chromium.org>
Commit-Queue: Charles . <charleszhao@chromium.org>