sirenia: Add dbus API for communicating with dugong
Add the xml file that defines the dbus API for communicating with dugong
from inside CrOS. Also add the dbus dependencies to the Cargo.toml and
the building of the API to build.rs. Built bindings can be found in
src/bindings.

BUG=b:162096925
TEST=built the package

Cq-Depend: chromium:2412812
Change-Id: Ica0931981de5165f3d3e518cee7a5c58848c3c00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2412901
Commit-Queue: Nicole Anderson-Au <nvaa@google.com>
Tested-by: Nicole Anderson-Au <nvaa@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>