cryptohome: Retrieve key for migrate without mount
The mount argument passed to Homedirs::Migrate is used for the sole
purpose of acquiring the id of the active key in the keyset. That can be
achieved, however, without actually mounting the vault. The added
out-argument return the index of the new key in case we already have the
mount active and need to update the key index on that.

In regards with the CL removing chaps migration path.
Even though mount populates the missing wrapped_chaps_key it is ok to
migrate without it being populated: the migrated key would also have
missing wrapped_chaps_key.
Thus, the following behaviour chains would both at the end trigger the
initial population of the missing wrapped_chaps_key on Mount in the same
manner:
(Update from R37 - Mount)
(Update from R37 - MigrateKey - Mount)

BUG=chromium:1124561
TEST=unittest, CQ && autotest

Change-Id: Icd9f9fe20f7ea4eb5a3667990775089d2867155f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2409747
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>