patchpanel: Simplify NeighborLinkMonitorTest
This patch refactors NeighborLinkMonitorTest a bit by storing the RTNL
listener registered by NeighborLinkMonitor, and thus we don't need to
call OnNeighborMessage() directly which is a private method. As a
result, no more FRIEND_TEST is needed and the test code can be
simplified.

Also fixes a typo and changes the usage of TaskEnvironment.

No actual test is changed by this patch.

BUG=None
TEST=cros_workon_make --board=$BOARD --test patchpanel

Change-Id: Icbcf624476a1f17d691a4acfb956afafb9bc1d65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2418491
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>