cryptohome: Set up WebAuthn secret during mounting
The WebAuthn secret is supposed to be unlocked only when the user
successfully authenticates themselves. However, u2fd needs to know
a public derivative (sha256 hash) of the secret when the user logs
in, so that u2fd can generate WebAuthn credentials using the hashed
secret. The unhashed secret will be used at WebAuthn GetAssertion
time.

This change adds the service.cc implementation of a DBus method
for u2fd to read the secret shortly after it's available. If it's not
read, the secret is cleared shortly.

A follow up change will add the user_data_auth implementation.

BUG=b:144861739
TEST=build and deploy cryptohome, login, verify the secret via logging.

Change-Id: I890d98816e080d2bd15df9ee4ba0fba98f3b1f7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2427360
Tested-by: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>