debugd: memory_test shouldn't be OOM unkillable
The command "memory_test" (run from crosh) has a -1000 OOM score.
That means it's not killable.  This seems unwise since, if something
happens after memory_test starts that chews up a bunch of memory, we'd
rather memory_test die than the system reboot.

This was seen in practice on two feedback reports.

Let's instead set things so that the memtester command is as killable
as possible.

NOTE: Do do this we use a tiny bit of code snarfed from the
oom_adj_tool.  The code is so little (and this is a standard Linux
API) that it didn't seem worth trying to share the code.

BUG=chromium:1129890
TEST=Start many copies of this and see one get oom killed.
TEST=Check /proc filesystem and see proper oom_score_adj.

Change-Id: I443f74b6b3c69d691a95e635793635f826d18636
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2419312
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>