cryptohome: Remove obsolete chaps key migration path.
This migration path facilitates upgrades from R37.
The tests for the path were dropped some time ago (http://shortn/_31PlU95WvN)
There is infenitisimal amount of user on pre R38 who are still eligible
for modern updates, and the fraction of those using chaps keys is
probably at least an order of magnitude lower.
The lack of the migration will invalidate the user's chaps database if
present if the user decides to finally update from R37 to a version
without the migration path support, but won't prevent loading vault.
They can fix the state by re-obtaining the certificates if the need
arise, but we deem it to be unlikely case.
The removal of the migration path unblocks an enormous amount of
refactoring and code cleanup oportunites.

See full justification in http://shortn/_hQzyfPge4J

BUG=none
TEST=unittest

Change-Id: I190a9168e3ad57d4fc0426051360f113c9275ad2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2424375
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>