crash: Remove --delete_crashes flag to sender.
Instead of using crash_sender to upload crashes from test runs, we'll
use a (to-be-introduced) crash_serializer binary.

Maintain .alreadyuploaded logic to guard against failure to remove
crashes, but change the action for .alreadyuploaded to kRemove, since
now the only way an .alreadyuploaded file could exist is from failure to
remove (as opposed to an explicit request not to remove).

BUG=chromium:1107005,chromium:1060019
TEST=CQ

Change-Id: I46a4e3a4fb8c6971156963ce26d5820daa553703
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2422089
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>