vm_tools: Add a reset lxd db override
Before we enable the feature by default it'd be nice to let users
opt-in, so they can recover from a broken container. This CL adds
plumbing for an override from Chrome (where it's wired to a Chrome flag)
and Tremplin, where it force-enables resetting even if it otherwise
wouldn't.

BUG=chromium:1057869
TEST=End-to-end. Broken container fails, toggle flag, now it starts

Disallow-Recycled-Builds: hatch-arc-r-cq
Change-Id: Ie6e527a81428cbcbd1418ee4f22714d9327c3668
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2425946
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>