lorgnette: Add cli flag to loop over all devices
If the new --all flag is passed, loop over all detected scanners and try
to request a scan from each one.  The same resolution and source are
passed to each.  Success or failure are tracked for each scanner and
printed at the end; scanning does not stop if one scanner fails.  This
is useful for confirming behavior changes on a group of scanners without
repeatedly running lorgnette_cli.

BUG=None
TEST=lorngette_cli --all with multiple scanners attached

Change-Id: I3b8eb013a606e0e9bbab65f07b5f25f63fd92b52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2429208
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>