REVERT-ME: power: Add support for SAR power level settings in Qualcomm modem
This is a temporary hack to support dynamic SAR in M87 for trogdor.
Whenever we detect a change in device mode or human presence,
the SAR power settings are adjusted accordingly using qmicli commands.
The following three power modes are supported
1 - Tablet + human presence
2 - Notebook + human presence
3 - No human presence
All of the code is under USE_TROGDOR_SSR_HACK
Refer go/cros-lte-sar-design-doc for details about the hack
and the proposed solution.

BUG=b:169701128
TEST=cros deploy ${DUT} power_manager && reboot

Cq-Depend: chromium:2446072
Change-Id: Ia0f15a5e9f1a90775f39fabc10c9a7474a18226b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2439495
Tested-by: Madhav . <madhavadas@google.com>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Madhav . <madhavadas@google.com>