crash: Check for 'done=1' meta key before payload.
We cannot rely on anything in the .meta file being valid before the
'done=1' key is written, but ChooseAction was checking whether the
'payload' key in the .meta file was empty or referred to a nonexistent
file _before_ checking that the .meta file was complete.

This leads to a race -- if crash_sender starts running while a crash is
being written, it could remove the meta file due to a missing payload,
even though crash_reporter would have later added the payload.

BUG=chromium:1132434
TEST=cros_workon_make --test --board=eve  crash-reporter
TEST=tast run $DUT crash.Sender*

Change-Id: I77593a26db4f7081e67198c15ca287e66ce71d59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2432518
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>