installer: Add mode for installing LVM stateful partition
Add a flag for setting up the stateful partition with an
LVM setup: create a physical volume and a volume group on
the stateful partition with a random name. Add thin logical
volumes for the unencrypted stateful partition and the
developer tools partition.

BUG=b:172344853
TEST=chromeos-install --lvm_stateful

Cq-Depend: chromium:2493230
Change-Id: I46e3ae956604b547b74b71d9c58ad38d32917941
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2434286
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>