shill: Report pin1 lock retries on dbus by default
The UI usually manipulates the PIN1 lock, irrespective of the actual lock. Thus, we report the PIN1 lock retries by default. If the lock is PUK lock, the UI will need to report PUK retries, therefore, we send the PUK retries only in this case. In the future, the UI will need to be aware of retry limits for all locks, irrespective of the actual lock set.(b/169615875)

BUG=b:162563902
TEST=FEATURES=test emerge-trogdor shill

Change-Id: I48936ad9ca18ab588a34e1617cda0a46913ddbb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2437102
Commit-Queue: Pavan Holla <pholla@google.com>
Tested-by: Pavan Holla <pholla@google.com>
Reviewed-by: Pavan Holla <pholla@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>