croslog: add a hack to ignore leading null chars in audit.log
There is a bug that null characters are unexpectedly inserted at the
begenning of some lines in audit.log for some reason (crbug.com/1132182).

We don't find a root cause of it, so that this CL adds a temporary
solution to handle these lines with a warning.

BUG=chromium:1132182
TEST=pass the added test.

Change-Id: If48ad8bb7d02107528bd83310ad3ab162f451b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2437558
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>