patchpanel: Add a fake class for Client
This patch adds a FakeClient class for making it easier for unit tests
in other packages. To do this, this patch also extracts an interface
from the original Client class, and renames the latter one to ClientImpl
and move it into an anonymous namespace in .cc file.

BUG=b:160158606
BUG=b:162194516
BUG=chromium:1135913
TEST=`FEATURES="test" emerge-${BOARD} patchpanel-client`
TEST=build and run patchpanel_client_fuzzer

Cq-Depend: chromium:2450136
Change-Id: I91cf7cc3e3057d940db9e65dcc4b97f8530ff8f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2438172
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>