shill: Add UMA metrics for comparing link monitors
This patch adds logic for comparing link monitors by comparing the
detection time diff for the same failure. To do this, we record the
detection time for the two link monitors in Device, which can only be
reset by either 1) connected event from patchpanel, or 2) another
service is selected. Also a timeout logic is used for the cases that the
failure is detected by one link monitor but not by another.

Chrome patch for the UMA metric: https://crrev.com/c/2462753

This patch can be reverted once we have finished the comparison.

BUG=b:162194516
TEST=manually set the neighbor entry to nud failed by `ip neighbor`
  command is able to trigger this metric.

Change-Id: Ib43ae86f3fe7bfd638fcbe1f69e8991a7a6bb39d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2439961
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>