metrics: add `include` to elide sandbox errors
The portage sandbox complains that we're walking 'up' to
`platform2/.git` when trying to `emerge` this package, so it can figure
out what files are needed for this package to build. This results in
sandbox errors. We can work around this with an explicit `include` in
the `[package]` section.

This also adds `rerun-if-changed` directives to the `build.rs` script,
so artifacts get properly invalidated if protos change (it's not
entirely clear to me that this would happen before).

BUG=chromium:1131513
TEST=`emerge-eve memd`

Change-Id: Ie8f9e490b62bd39ba707e01f73e3619faef636f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2441126
Reviewed-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Auto-Submit: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>