crash: Put exec, collector in failure reports.
Previously, in https://crrev.com/c/2136289, we stopped reporting these
failures using the original exec_name and collector. This still is
reasonable, because reporting that way masks the actual problem.

However, it is not ideal in that it obscures which execs/collectors are
having issues. Add this information back to the metadata so that we can
determine whether any collectors or exec names have a disproportionate
rate of failures.

BUG=chromium:1039296,chromium:1124382
TEST=test_that $DUT logging_UserCrash
TEST=tast run $DUT platform.UserCrash.*

Change-Id: Idfd39d1d6029b9b2d06489a75300827c3b1f9eae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2441180
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>