cryptohome: Have a specific error for vault recreation trigger.
Having unspecific MOUNT_ERROR_FATAL is too abstract for people to
realize it may cause a destructive behaviour. Here we introduce a
dedicated error to trigger the re-creation if all keys are unparsable or
broken in another way. This change reduces the amount of cases
triggering the recreation - e.g. just missing output vk structure is an
error but, but not a reason for vault re-creation.

BUG=chromium:1124561,chromium:1132026
TEST=unittest

Change-Id: Id3b9393df40f88abb5c4781f22f9fe81f65c6f0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2439970
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>