crash: Refactor util::Sender.
As the first step of writing the crash serializer, pull out
functionality that will be common to both the sender and the serializer
into a new base class that both will inherit.

This is a pure refactor and should _not_ have any functional changes.

BUG=chromium:1107005
TEST=unit tests; tast run $DUT crash.Sender*
TEST=test_that $DUT logging_UserCrash logging_CrashSender

Change-Id: I20720ffe31b14d03441919333b6927ccaa5b8430
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2433051
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>