cryptohome: remove is_pristine from CreateTrackedDirectory
The flag let the code know whether the directory is freshly created.
CreateTrackedDirectory tries to rectify a mismatch state on the
NOT-freshly-created cases. There is no reason not to do the same in the
case of the freshly created vault - it should be pretty much a no-op as
no dirs shall exist yet. That removes one of only two use cases of the
is_pristine check in the mount process, and the other will be removed in
the following CL.

There was a case in 2011 where tracked directories where introduced and
downloads was tracked; it needed to be saved when the directory was
created in clear in the ecrypfs vault
(see https://codereview.chromium.org/6544019). The migration was completed
and code removed in 2012 with https://gerrit.chromium.org/gerrit/18824.

BUG=chromium:1124561
TEST=unittest

Change-Id: I34d57f4a3c8eee381935669e56ad5f27a66673eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2444609
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>