diagnostics: remove commas from marketing_name
The marketing_name field of SystemInfo sometimes had commas,
which caused the platform.CrosHealthdProbeSystemInfo tast test
to fail. Replace the commas with a forward slash. This keeps
the same information, but allows the tast test to reason about
it.

BUG=chromium:1134667
TEST=emerge-nami diagnostics
platform.CrosHealthdProbeSystemInfo in the CQ should test more.

Change-Id: I6df4ae08759bfbee33b16d1f3ad31024a9b55a2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2445465
Tested-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Trent Begin <tbegin@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>