login: Re-purpose Login.MountNamespaceCreationSuccess
Move the namespace creation success metric from test namespace creation
to actual namespace creation.
We've collected enough data with the test namespace creation and don't
need to have a metric reporter with that any more.
But we also want to have metrics reported with the actual namespace
creation once it is enabled for regular sessions. This CL moves the
metrics to actual namespace creation block. That will stop the namespace
creation success metrics temporarily. But it'll report the success rate
for actual namespace creation once it is enabled.

BUG=chromium:1072159
TEST=Compiles.

Change-Id: I14f76d547655d7d618ca2edbf68f74f96fc3d544
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2449166
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>