adbd: Automatically restart service
The main reason to respawn this service is to deal with changes
from USB cable. The service so far is designed to quit when the
cable is disconnected and relies on the upstart to restart it.

The service can quit under other error conditions as well, so it
makes sense to relaunch the service to recover.

BUG=chromium:1128701
TEST=ADB works again after cable is removed and then attached to
  the eve device (eve-arc-r).

Change-Id: If8583a75cc17a119248b84387da11cb9fc1832f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2450453
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Jianxun Zhang <jianxun.zhang@intel.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>