croslog: Send croslog's errors and warnings to syslog
This CL adds an option to send logs from croslog command to
syslog.  If the stdout is not tty, this option is automatically
enabled. Or the user can manually specify the command line option.

This is introduced to find a clue of crbug.com/1132182

BUG=chromium:1132182
TEST=manually tested

Change-Id: If67a3474e0a0e4b2fa66479cf8cfedc288cf0af4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2449517
Reviewed-by: Yuta Hijikata <ythjkt@chromium.org>
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>