vm_tools: Make garcon conformant to XDB base spec
The base directory spec defines (among other things) which paths to
look in for .desktop files. Previously it was not conformant to that
spec, specifically it did not have the correct defaults and missed
searching the user's data dir.

This change makes it conform to the spec.

A similar problem with icon searching was also fixed by having the icon
finder call the .desktop finder.

BUG=b:162562622
TEST=modify ~/.profile and check garcon's logs

Change-Id: Id864001183aa7bb9454828366c2b9c6f12553dcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2454473
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>