shill: vpn: use callback for tunnel creation
This patch adds a callback to DeviceInfo::CreateTunnelInterface
which would be triggered when link is ready, to substitute the
VPNDriver::ClaimInterface iteration used to pass VirtualDevice
ownership.

Currently the callbacks still point back to the VPNDrivers but
will later be moved to VPNServices. OnDeviceInfoAvailable is
retained in this patch only to track ArcBridge ready event, which
will also be migrated in the latter patches.

BUG=b:170478571
BUG=chromium:1026648
TEST=unit; manually verify openvpn and chrome 3rd party vpn

Change-Id: I035ce858d1e6a003536634fa90915c5d6b45600f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2463265
Tested-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>