cryptohome: minimize usage of keyset index in Mount
Remove passing raw keyset index from most of the interfaces in Mount.
We still need the index in places for key managements, but at least we
are mostly not passing it around. Usage of VaultKeyset::legacy_index is
valid because we always populate it on Load path within
Homedirs:LoadVaultKeysetForUser

BUG=chromium:1124561
TEST=unittest

Change-Id: I6e606ed731ce33aa568264acb628e4b79ad7d615
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2465549
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>