vm_tools: seneschal: ignore missing dst for unshare
Previously seneschal would return early with error in UnsharePath if
|dst| path does not exist, but it is expected that |dst| may not exist
when a path is unshared due to a volume unmount, or if a path is
unshared from the settings page when a volume is not mounted. In such
cases, we should continue processing to clean up empty mount points and
return success.

BUG=chromium:1132707
BUG=chromium:1133621
TEST=tast run -var=keepState=true $DUT crostini.Share*.artifact

Change-Id: Id6629fb52568c39e4d9b3ac44946ee47b3e1fc81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2465892
Tested-by: Joel Hockey <joelhockey@chromium.org>
Auto-Submit: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Josh Simmons <simmonsjosh@google.com>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>