croslog: Ignore invalid boot entry
There may be an invalid boot entries which doesn't stored in an
incremental order. This may happened when the system clock is
changed backed or the system clock was wrong.

Previously, it was treated as a fatal error. This CL makes it with
just a warning and an invalid one is just ignored.

BUG=chromium:1137400
TEST=pass an added test

Change-Id: I98f2ceb7323a0e9bb19cbb639add1d42b80e75bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2467301
Reviewed-by: Yuta Hijikata <ythjkt@chromium.org>
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>