cryptohome: Do not create Mount in out-of-session Migrate
Homedirs::Migrate earlier required Mount as an argument, since it
doesn't need it any longer, we shall stop creating the placeholder mount
object for out-of-session Migration as we do not need to update
in-memory creds in the case. Because of that, we do not need any longer
a separate test for unmounted case as it is pretty much follows the same
logic.

BUG=chromium:1124561
TEST=unittest

Change-Id: I4a1f50865d8afea98b1a370f98989fc8aff41cdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2470227
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>