cryptohome: Add new UserSession and use it UserDataAuth
New user session will slowly get all the logic from UserSession and will
encapsulate some mount code path routing logic. At this time, however,
it is a simple wrapper around Mount, so we have a minimal change of all
the references to the object. This CL changes the references in
userdataauth and the following Cl will do the same for service. Pretty
much all of the changes are mechanical replacements of references to
Mount with UserSession::GetMount. There are a lot of place where the
code could be improved, but to make the review simple, we are not
changing any logic in this CL.

BUG=chromium:1124561
TEST=unittest

Change-Id: I1c6878f2b8401ef97b670f7747bb08048f6bf4c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2470229
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>