cryptohome: User new UserSession in Service
New user session will slowly get all the logic from UserSession and will
encapsulate some mount code path routing logic. At this time, however,
it is a simple wrapper around Mount, so we have a minimal change of all
the references to the object. This CL changes the references in
service and the previous CL did the same for userdataauth. Pretty
much all of the changes are mechanical replacements of references to
Mount with UserSession::GetMount except for a few peculiar places.
There are a lot of place where the code could be improved, but to make the
review simple, we are not changing any logic in this CL.

BUG=chromium:1124561
TEST=unittest

Change-Id: I60379bd2f4493c1c420b383211c70aa3161da95b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2470230
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>