cryptohome: Move logic LegacyUserSession->UserSession
The chain of CLs this one is the apex of is target to remove one of the
major consumers of credentials in Mount object by inverting the
dependency between UserSession and Mount. The previous CLs established
the platform for the change, by renaming the old UserSession to
LegacyUserSession, instantiating a proxy object and updating all the
references to Mount and LegacyUserSessuion on service/userdataauth
level. This CL finishes the work by actually moving the logic from of
the old UserSession to the new one and removes the references to the
UserSession from the Mount object. This opens the path to make Mount
object completely credentials agnostic.

BUG=chromium:1124561
TEST=unittest

Change-Id: I5cde91ec3fef0d4ece30c79bcd3b182bb4e5d405
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2470233
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>