ml: Move tclib model file installation to tclib's ebuild.
We used to put the model installation in mlservice's ebuild because we
thought the unit test can not see the model files in the standard
location. It turns it can. So it is better to install the model files
in its own ebuild.

Now we do not need to load the model file in
`MachineLearningServiceImpl` and can put the model initialization all
in "text_classifier_impl.cc"

Also removed a unit test on testing non-exist model file for tclib
because it seems there is no such tests for all the other models.

BUG=none
TEST=FEATURES=test emerge-nocturne ml

Cq-Depend: 2472978
Change-Id: I4ac44446b99815705e2c82dc543b41170c7f7fb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2473178
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Honglin Yu <honglinyu@chromium.org>
Tested-by: Honglin Yu <honglinyu@chromium.org>