timberslide: fix buffer overflow
The `buffer` array is used as a c-style string in ProcessLogBuffer. If
ReadAtCurrentPosNoBestEffort() reads 4096 bytes into the buffer, the
missing '\0' will cause istringstream accessing the buffer out of bound.

Change the c-style string to c++ string, and construct it with the
correct length information.

Also remove the memset() since we don't need trailing '\0' anymore.

BUG=b:167475144
TEST=1) no garbage in ec log
     2) assert length of string <= 4096

Change-Id: I43da3c1e28df11c4ad9c474687c6bfbd453485b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2478842
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>