cryptohome: remove broken reset_seed population on Mount
The code path for repopulation of the reset_seed on Mount is currently
broken, because the reset_seed is unlikely to be re-saved then: the only
condition that it will be re-saved is if chaps_key was also missing, but
that is only possible if it is the first mount with the key since
upgrade from R37. That, however, doesn't matter, since the reset_seed is
populated whenever a new keyset is added - which is going to happen
before LE creds are stored for the user. (see Homedirs::AddKeyset).
Once the chain [1] is submitted, all of the re-encryption cases can be
trivially moved  to Homedirs::LoadVaultKeysetForUser.

[1] https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2465546

BUG=chromium:1124561,chromium:1138471
TEST=unittest

Disallow-Recycled-Builds: harch-arc-r-cq
Change-Id: I32238e84d12ff85c14983b40352f3bf95d9bf836
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2478703
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>